requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Babel==2.15.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
cryptography==43.0.0
cssselect2==0.7.0
Django==5.0.7
django-crispy-forms==2.2
django-environ==0.11.2
django-filter==24.2
django-forms-dynamic==1.0.0
django-money==3.5.2
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-tables2==2.7.0
django-widget-tweaks==1.5.0
fonttools==4.53.1
html5lib==1.1
idna==3.7
lxml==5.2.2
mysqlclient==2.2.4
numpy==2.0.0
oscrypto==1.3.0
pango==0.0.1
phonenumbers==8.13.40
pillow==10.4.0
py-moneyed==3.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==4.3.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
